refactor: 💡 custom hooks, rm unneccessary vars, params
