+++ src/slashMenu/index.tsx
@@ -1,6 +1,5 @@
 export * from "./SlashMenuPlugin";
 export {
-  // TOGGLE_SLASH_MENU,
   REMOVE_LAST_FILTER_CHARACTER,
   ADD_FILTER_CHARACTER,
   RESET_STATE,
