+++ src/slashMenu/utils.ts
@@ -1,5 +1,26 @@
 import { ItemId, MenuElement, BaseSlashMenuState, SubMenu } from "./types";
 
+export function getDOMRangeRect(
+  nativeSelection: Selection,
+  rootElement: HTMLElement
+): DOMRect {
+  const domRange = nativeSelection.getRangeAt(0);
+
+  let rect;
+
+  if (nativeSelection.anchorNode === rootElement) {
+    let inner = rootElement;
+    while (inner.firstElementChild != null) {
+      inner = inner.firstElementChild as HTMLElement;
+    }
+    rect = inner.getBoundingClientRect();
+  } else {
+    rect = domRange.getBoundingClientRect();
+  }
+
+  return rect;
+}
+
 export const getElementIds = (item: MenuElement): ItemId[] => {
   if (item.type === "submenu")
     return [
